Corrected amount of initial cards to 10 implemented drawing cards if the pick is 3 or more like original game

+++ CardWebHooks/Cards/Deck.cs
@@ -107,7 +107,7 @@ namespace CardWebSocks.Cards
         public string Text { get; set; }
 
         [JsonProperty("pick")]
-        public long Pick { get; set; }
+        public int Pick { get; set; }
 
         public BlackCard(string text, int pick)
         {
